+++ Editor/AnimatorParserV2/AnimatorParser.cs
@@ -550,7 +550,7 @@ namespace Anatawa12.AvatarOptimizer.AnimatorParsersV2
             }
         }
 
-        public (AnimatorController, IReadOnlyDictionary<AnimationClip, AnimationClip>) GetControllerAndOverrides(
+        public static (AnimatorController, IReadOnlyDictionary<AnimationClip, AnimationClip>) GetControllerAndOverrides(
             RuntimeAnimatorController runtimeController)
         {
             if (runtimeController is AnimatorController originalController)