+++ Editor/ContextExtensions.cs
@@ -25,7 +25,7 @@ namespace Anatawa12.AvatarOptimizer
         public static MeshInfo2 GetMeshInfoFor([NotNull] this BuildContext context, SkinnedMeshRenderer renderer) =>
             context.GetHolder().GetMeshInfoFor(renderer);
 
-        private static ObjectMappingBuilder GetMappingBuilder([NotNull] this BuildContext context)
+        public static ObjectMappingBuilder GetMappingBuilder([NotNull] this BuildContext context)
         {
             if (context == null) throw new ArgumentNullException(nameof(context));
             return context.Extension<ObjectMappingContext>().MappingBuilder;