+++ Test~/AnimatorOptimizer/EntryExitToBlendTree/EntryExitToBlendTreeTest.cs
@@ -1,0 +1,18 @@
+using Anatawa12.AvatarOptimizer.Processors.AnimatorOptimizer;
+using NUnit.Framework;
+
+namespace Anatawa12.AvatarOptimizer.Test.AnimatorOptimizer
+{
+    public class EntryExitToBlendTreeTest : AnimatorOptimizerTestBase
+    {
+        [Test]
+        public void GestureConvertibleSimple()
+        {
+            var controller = LoadCloneAnimatorController("GestureConvertibleSimple");
+            controller.name = "GestureConvertibleSimple.converted";
+            EntryExitToBlendTree.Execute(new AOAnimatorController(controller));
+            var except = LoadAnimatorController("GestureConvertibleSimple.converted");
+            RecursiveCheckEquals(except, controller);
+        }
+    }
+}