+++ Test~/AnimatorParser/AnimatorTest.cs
@@ -247,11 +247,10 @@ namespace Anatawa12.AvatarOptimizer.Test.AnimatorParserTest
         [Test]
         public void TestOneLayerOverrides()
         {
-            var parser = new AnimatorParser(true);
             var controller = TestUtils.GetAssetAt<RuntimeAnimatorController>("AnimatorParser/OneLayerOverrideController.overrideController");
             var animate0To100 = TestUtils.GetAssetAt<AnimationClip>("AnimatorParser/Animate0To100.anim");
             var animate1To100 = TestUtils.GetAssetAt<AnimationClip>("AnimatorParser/Animate1To100.anim");
-            var (original, mapping) = parser.GetControllerAndOverrides(controller);
+            var (original, mapping) = AnimatorParser.GetControllerAndOverrides(controller);
 
             Assert.That(original, Is.EqualTo(_controller));
             Assert.That(mapping, Is.EquivalentTo(new []
@@ -263,13 +262,12 @@ namespace Anatawa12.AvatarOptimizer.Test.AnimatorParserTest
         [Test]
         public void TestTwoLayerOverrides()
         {
-            var parser = new AnimatorParser(true);
             var controller = TestUtils.GetAssetAt<RuntimeAnimatorController>("AnimatorParser/TwoLayerOverrideController.overrideController");
             var animate0To100 = TestUtils.GetAssetAt<AnimationClip>("AnimatorParser/Animate0To100.anim");
             var animate1To100 = TestUtils.GetAssetAt<AnimationClip>("AnimatorParser/Animate1To100.anim");
             var animate1ToVariable = TestUtils.GetAssetAt<AnimationClip>("AnimatorParser/Animate1ToVariable.anim");
             var animate2ToVariable = TestUtils.GetAssetAt<AnimationClip>("AnimatorParser/Animate2ToVariable.anim");
-            var (original, mapping) = parser.GetControllerAndOverrides(controller);
+            var (original, mapping) = AnimatorParser.GetControllerAndOverrides(controller);
 
             Assert.That(original, Is.EqualTo(_controller));
             Assert.That(mapping, Is.EquivalentTo(new []